Master thesis base version
This was quickly pulled from the original repository to remove parts of the source that could not be included. At this stage it is not as complete as it should be, and no history of the development is present, but it is the stage of the project from when I delivered my master thesis. Looking forward to work further with the project, especially to delve deeper into dash app development, and making the tool project agnostic.