[hopefully] ready for use
most bugs should be gone now. Might be some small ones I've overlooked. But it seems to be working with a correct metadata file and a database to setup backend, frontend, and host the iof app on local computer. Next is to write a detailed guide on these steps, although it should be pretty straightforward