ran black on iof_app and changed 3D render of cages
python-black of iof_app.py code. Haven't manually gone through to change potential weird parts. Should be fine for the most part. Also reduced the complexity of the 3D rendering of cages drastically. The app doesn't take as long to load, and the 3D plots are slightly less laggy as a result. And the render looks better in my opinion. I've also made a script to convert themlabiotel databases to iof, but this is not included in the git repo for now.